DchLexer: Reduce number of whitespace tokens
Previously, the lexer would generally emit many single-whitespace
tokens.  With this change, we identify some key points where the
lexer can "trivially" emit a multi-character whitespace token
(instead of multiple single-character whitespace tokens).

The problem arose with "start of line" semantics where we use
whitespace to determine which "state" the lexer should be in.
Fortunately, this rule is basically the same no matter where we are in
the lexer, which enables us to special case the patterns "NL+ SP SP"
and "NL+ SP" to a single whitespace token (along with jumping directly
to the relevant state).

Signed-off-by: Niels Thykier <niels@thykier.net>