Parse Build-Profiles directly in Deb822Lexer
The current way of parsing deb822 fields generically is not working
very well (injected languages loses have no context from their
"host").  This commit rewrites the deb822 lexer to have context based
parsing by leveraging the underlying language.  This feels like a hack
but it is the best solution I could come up that was the "least of
amount of fighting with IntelliJ".

Signed-off-by: Niels Thykier <niels@thykier.net>